adds cli options for scraper and some cleanups
