+++ bin/scrape-docs/Cargo.toml
@@ -1,14 +1,15 @@
 [package]
-name = "scrape-docs"
-version = "0.1.0"
 authors = ["Lonami Exo <totufals@hotmail.com>"]
 edition = "2018"
+name = "scrape-docs"
+version = "0.1.0"
 
 [dependencies]
-reqwest = "0.11.3"
-tokio = { version = "1.5.0", features = ["full"] }
-select = "0.5.0"
-serde = { version = "1.0.125", features = ["derive"] }
-serde_json = "1.0.64"
+clap = {version = "3.1.8", features = ["derive"]}
 futures = "0.3.14"
-regex="1.5.5"
+regex = "1.5.5"
+reqwest = {version = "0.11.3", features = ["cookies"]}
+select = "0.5.0"
+serde = {version = "1.0.125", features = ["derive"]}
+serde_json = "1.0.64"
+tokio = {version = "1.5.0", features = ["full"]}
